+++ src/app/inobeta-ui/ui/kai-filter/filter.component.ts
@@ -19,11 +19,11 @@ import { IB_FILTER } from "./tokens";
   template: `
     <section
       class="ib-filter"
-      [class.ib-filter__hide]="hideFilters"
+      [class.ib-filter--hidden]="hideFilters"
       [attr.aria-hidden]="hideFilters"
     >
       <ng-content select="ib-search-bar"></ng-content>
-      <section #list class="ib-filter-list">
+      <section #list class="ib-filter__list">
         <mat-icon *ngIf="list.children.length > 1">filter_list</mat-icon>
         <ng-content></ng-content>
       </section>
@@ -69,7 +69,7 @@ export class IbFilter {
     });
   }
   @Output() ibFilterUpdated = new EventEmitter<IbFilterSyntax>();
-  @Output() ibRawFilterUpdated = new EventEmitter<Record<string, any>>();
+  @Output() ibQueryUpdated = new EventEmitter<Record<string, any>>();
 
   /** @ignore */
   form: FormGroup = new FormGroup<Record<string, any>>({});
@@ -88,13 +88,24 @@ export class IbFilter {
     this.rawFilter = this.form.getRawValue();
     this.filter = this.buildFilter();
     this.ibFilterUpdated.emit(this.filter);
-    this.ibRawFilterUpdated.emit(this.rawFilter);
+    this.ibQueryUpdated.emit(this.toQuery());
   }
 
   reset() {
     this.form.reset();
     this.form.patchValue(this.initialRawValue);
     this.update();
+  }
+
+  toQuery() {
+    let output = {};
+    const filters = this.filters.toArray();
+
+    for (const filter of filters) {
+      output[filter.name] = filter.toQuery();
+    }
+
+    return output;
   }
 
   /** @ignore */